\name{Environment}
\alias{Environment}
\docType{data}
\title{Attitude to the Environment}
\description{
 This data set comes from the Environment section of the 1990 British Social Attitudes Survey (Brook et al., 1991). 
 A sample of 291 responded to the questions below:
}
\format{
All of the below items were measured on a three-group scale with response categories "very concerned", 
"slightly concerned" and "not very concerned":
  \describe{
    \item{LeadPetrol}{Lead from petrol.}
    \item{RiverSea}{River and sea pollution.}
    \item{RadioWaste}{Transport and storage of radioactive waste.}
    \item{AirPollution}{Air pollution.}
    \item{Chemicals}{Transport and disposal of poisonous chemicals.}
    \item{Nuclear}{Risks from nuclear power station.}
  }
}
\references{
  Bartholomew, D., Steel, F., Moustaki, I. and Galbraith, J. (2002) \emph{The Analysis and Interpretation of 
  Multivariate Data for Social Scientists}. London: Chapman and Hall.
  
  Brook, L., Taylor, B. and Prior, G. (1991) \emph{British Social Attitudes, 1990, Survey}. London: SCPR.

}
\examples{

## Descriptive statistics for Environment data
descript(Environment)

}
\keyword{datasets}
