\name{Mobility}
\alias{Mobility}
\docType{data}
\title{ Women's Mobility }
\description{
  A rural subsample of 8445 women from the Bangladesh Fertility Survey of 1989.
}
\format{
The dimension of interest is women's mobility of social freedom. Women were asked
whether they could engage in the following activities alone (1 = yes, 0 = no):  
  \describe{
    \item{Item 1}{Go to any part of the village/town/city.}
    \item{Item 2}{Go outside the village/town/city.}
    \item{Item 3}{Talk to a man you do not know.}
    \item{Item 4}{Go to a cinema/cultural show.}
    \item{Item 5}{Go shopping.}
    \item{Item 6}{Go to a cooperative/mothers' club/other club.}
    \item{Item 7}{Attend a political meeting.}
    \item{Item 8}{Go to a health centre/hospital.}
  }
}
\source{
  Bangladesh Fertility Survey of 1989 (Huq and Cleland, 1990).
}
\references{
  Bartholomew, D., Steel, F., Moustaki, I. and Galbraith, J. (2002) \emph{The Analysis and Interpretation of 
  Multivariate Data for Social Scientists}. London: Chapman and Hall.
  
  Huq, N. and Cleland, J. (1990) \emph{Bangladesh Fertility Survey, 1989}. Dhaka: National
  Institute of Population Research and Training (NIPORT).
}
\examples{

## Descriptive statistics for Mobility data
descript(Mobility)

}
\keyword{datasets}
