\name{Science}
\alias{Science}
\docType{data}
\title{Attitude to Science and Technology}
\description{
 This data set comes from the Consumer Protection and Perceptions of Science and Technology section of the 
 1992 Euro-Barometer Survey (Karlheinz and Melich, 1992) based on a sample from Great Britain. The questions 
 asked are given below:
}
\format{
All of the below items were measured on a four-group scale with response categories "strongly disagree", 
"disagree to some extent", "agree to some extent" and "strongly agree":
  \describe{
    \item{Comfort}{Science and technology are making our lives healthier, easier and more comfortable.}
    \item{Environment}{Scientific and technological research cannot play an important role in protecting
        the environment and repairing it.}
    \item{Work}{The application of science and new technology will make work more interesting.}
    \item{Future}{Thanks to science and technology, there will be more opportunities for the future generations.}
    \item{Technology}{New technology does not depend on basic scientific research.}
    \item{Industry}{Scientific and technological research do not play an important role in industrial development.}
    \item{Benefit}{The benefits of science are greater than any harmful effect it may have.}
  }
}
\references{
  Bartholomew, D., Steel, F., Moustaki, I. and Galbraith, J. (2002) \emph{The Analysis and Interpretation of 
  Multivariate Data for Social Scientists}. London: Chapman and Hall.
  
  Karlheinz, R. and Melich, A. (1992) Euro-Barometer 38.1: \emph{Consumer Protection and Perceptions of Science 
  and Technology}. INRA (Europe), Brussels. [computer file]

}
\examples{

## Descriptive statistics for Science data
descript(Science)

}
\keyword{datasets}
